git-svn-id: file://localhost/home/tolik/svnrepo/tags/latest_tested_snapshot@1787 c5418bc3-7546-4aae-8bad-fa6f2f3e53e6

+++ opencv/src/cxcore/cxmatrix.cpp
@@ -1217,7 +1217,7 @@ void MatND::create(const Vector<int>& _sizes, int _type)
         dim[i].step = total;
         total1 = (int64)total*sz;
         CV_Assert( sz > 0 );
-        if( total1 != (size_t)total1 )
+        if( (uint64)total1 != (size_t)total1 )
             CV_Error( CV_StsOutOfRange, "The total matrix size does not fit to \"size_t\" type" );
         total = (size_t)total1;
     }
